Decouple channel-sets
